+++ apps/client/src/app/pages/profile/public-profile/public-profile.component.ts
@@ -13,8 +13,8 @@ import { CraftingRotation } from '../../../model/other/crafting-rotation';
 import { CraftingRotationService } from '../../../core/database/crafting-rotation/crafting-rotation.service';
 import { FirestoreListStorage } from '../../../core/database/storage/list/firestore-list-storage';
 import { Apollo } from 'apollo-angular';
-import gql from 'graphql-tag';
 import { TranslateService } from '@ngx-translate/core';
+import * as semver from 'semver';
 
 @Component({
   selector: 'app-public-profile',
@@ -33,7 +33,7 @@ export class PublicProfileComponent {
 
   public communityRotations$: Observable<CraftingRotation[]>;
 
-  public fishingRankings$: Observable<any>;
+  public fishingRankings$: Observable<any> = EMPTY;
 
   public notFound = false;
 
@@ -63,7 +63,18 @@ export class PublicProfileComponent {
     );
     this.communityLists$ = userId$.pipe(
       switchMap(userId => {
-        return this.listsService.getUserCommunityLists(userId);
+        return this.listsService.getUserCommunityLists(userId).pipe(
+          map(lists => {
+            return lists.sort((a, b) => {
+              if (semver.gt(a.version, b.version)) {
+                return -1;
+              } else if (semver.gt(b.version, a.version)) {
+                return 1;
+              }
+              return b.forks - a.forks;
+            });
+          })
+        );
       }),
       shareReplay(1)
     );
@@ -73,48 +84,48 @@ export class PublicProfileComponent {
       }),
       shareReplay(1)
     );
-    this.fishingRankings$ = this.user$.pipe(
-      switchMap(user => {
-        return this.apollo.query<any>({
-          query: gql`query userRankings {
-            user_rankings:fishingresults(where:{userId: {_eq: "${user.$key}"}, ranking: { rank: {_lte: 3}}, itemId: {_gt: 0}}) {
-                size,
-                date,
-                itemId,
-                baitId,
-                ranking {
-                  rank
-                }
-            }
-          }`,
-          fetchPolicy: 'no-cache'
-        });
-      }),
-      map(result => {
-        const dataDisplay = result.data.user_rankings
-          .reduce((display, row) => {
-            const displayRow = display.find(dRow => {
-              return dRow.itemId === row.itemId;
-            });
-            if (displayRow && displayRow.ranking.rank <= row.ranking.rank) {
-              display = display.filter(dRow => dRow.itemId !== row.itemId);
-            }
-            display.push(row);
-            return display;
-          }, [])
-          .sort((a, b) => {
-            return a.ranking.rank - b.ranking.rank;
-          });
-        return {
-          summary: {
-            1: dataDisplay.filter(row => row.ranking.rank === 1).length,
-            2: dataDisplay.filter(row => row.ranking.rank === 2).length,
-            3: dataDisplay.filter(row => row.ranking.rank === 3).length
-          },
-          data: dataDisplay
-        };
-      })
-    );
+    // this.fishingRankings$ = this.user$.pipe(
+    //   switchMap(user => {
+    //     return this.apollo.query<any>({
+    //       query: gql`query userRankings {
+    //         user_rankings:fishingresults(where:{userId: {_eq: "${user.$key}"}, ranking: { rank: {_lte: 3}}, itemId: {_gt: 0}}) {
+    //             size,
+    //             date,
+    //             itemId,
+    //             baitId,
+    //             ranking {
+    //               rank
+    //             }
+    //         }
+    //       }`,
+    //       fetchPolicy: 'no-cache'
+    //     });
+    //   }),
+    //   map(result => {
+    //     const dataDisplay = result.data.user_rankings
+    //       .reduce((display, row) => {
+    //         const displayRow = display.find(dRow => {
+    //           return dRow.itemId === row.itemId;
+    //         });
+    //         if (displayRow && displayRow.ranking.rank <= row.ranking.rank) {
+    //           display = display.filter(dRow => dRow.itemId !== row.itemId);
+    //         }
+    //         display.push(row);
+    //         return display;
+    //       }, [])
+    //       .sort((a, b) => {
+    //         return a.ranking.rank - b.ranking.rank;
+    //       });
+    //     return {
+    //       summary: {
+    //         1: dataDisplay.filter(row => row.ranking.rank === 1).length,
+    //         2: dataDisplay.filter(row => row.ranking.rank === 2).length,
+    //         3: dataDisplay.filter(row => row.ranking.rank === 3).length
+    //       },
+    //       data: dataDisplay
+    //     };
+    //   })
+    // );
   }
 
 }